dags/requirements.txt
apache-airflow-providers-postgres==6.1.2
apache-airflow-providers-amazon==9.6.0
sagemaker==2.243.1
psycopg2-binary==2.9.10
apache-airflow[statsd]
flake8==7.2.0
black==25.1.0
pytest==8.3.5
pendulum==3.0.0
PyYAML==6.0.2